Refactorization End
Now program is in the similar state like the previous one. Refactorizated program has better structure, code is more readable and gives more opportunities to add new features easly. Structure of the level txt files is simplified and current format will not be working for the old code. Now I will focus on create LevelEditor and Menu of the game