Coin add button
I have added button for coin adding. It is possible to drag and drop it similar to the enemies. I have changed a displaying order, now when something is selected it goes to the front of the screen. The issue is that is possible to select multiple objects with one click when objects are overlapping and mouse coordinates belong to the rect area of the multiple overlapping objects.